frobshop/requirements.txt
asgiref==3.7.2
Babel==2.13.1
Django==3.2.14
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-q
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==20.1.0
Flask-QRcode
gunicorn
phonenumbers==8.13.26
Pillow==9.2.0
psycopg2-binary==2.9.3
purl==1.6
python-dateutil==2.8.2
pytz==2023.3.post1
qrcode==6.1
redis
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
typing_extensions==4.9.0
